move: useClickOutside 파일을 clientStateHooks 폴더로 이동
